+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- Interaction methods are faster and safer now. Update your scripts!
 
 ## [v1.1.0] - 2016-06-03
 ### Fixed