+++ CHANGELOG.md
@@ -10,7 +10,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Messages to the tracked object and all children:
   - `ViveGripHighlightStart` and `ViveGripHighlightStop`
   - `ViveGripGrabStart` and `ViveGripGrabStop`
-- Example right hand
+- Example hands and documentation
+
+### Fixed
+- Honouring the `.enabled` flag on grabbable and interactable scripts
 
 ## [v1.1.0] - 2016-06-03
 ### Fixed