+++ CHANGELOG.md
@@ -3,8 +3,11 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+
+## [v1.2.0] - 2016-06-11
 ### Changed
-- `ViveGripInteractionHeld` is now `ViveGripInteractionStop` (*update your scripts!*)
+- `ViveGripInteractionHeld` is now `ViveGripInteractionStop` (**update your scripts!**)
 - All `ViveGripXStart` and `ViveGripXStop` scripts pass in the calling ViveGrip_GripPoint object
 
 ### Added
@@ -47,7 +50,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Demo scene
 - Documentation
 
-[Unreleased]: https://github.com/JScott/ViveGrip/compare/v1.1.0...HEAD
+[Unreleased]: https://github.com/JScott/ViveGrip/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/JScott/ViveGrip/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/JScott/ViveGrip/compare/v1.0.2...v1.1.0
 [1.0.2]: https://github.com/JScott/ViveGrip/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/JScott/ViveGrip/compare/v1.0.0...v1.0.1